Introduces a more formalized `Dependency` approach
Here I'm formalizing the idea of the `Dependency` task parameters, with
a convention that dependency subclasses are callable to return the
actual value they will give the to the task at runtime.  They are given
the current docket, worker, and execution as context, but they can
return any value.

I've introduced `CurrentDocket` and `CurrentWorker`, and I've reworked
`Retry` to this framework.