Adds a chaos testing suite
This suite gives us a way to test some of the resiliency features of
`docket` in ways that are challenging to the unit test suite.  Here we
can run many processes, crash them, even restart the redis server in
order to see how the system behaves.

As a fun side note, I got some extra help with chaos testing here
because on my system I would often overload the proxy started by
`testcontainers` and cause errors connecting to redis throughout.  This
gave me a lot of real-world examples of places to handle errors in a
safer way.

This also required implementing `Worker.run_forever`, which is now in
place, if a little kludgey.

Closes #16
Part of #8