Adds OpenTelemetry spans for executed tasks, and links to the scheduler (#20)
* Adds OpenTelemetry spans for executed tasks, and links to the scheduler

This introduces OpenTelemetry tracing for `docket`, including a test
environment using [SigNoz](https://signoz.io/) to visualize the traces
and logs.

This also includes some improvements to the chaos testing suite in the
hopes of making it a little more stable in CI.  If nothing else, it will
have a timeout now.

Closes #6
Closes #19

* Delete .vscode/settings.json