Adds metrics for all activities, in OpenTelemetry and Prometheus
Here we use the OpenTelemetry metric objects to implement docket actions
like adding, cancelling, starting, or retrying tasks.  Using this
approach with the OpenTelemetry metrics gives us Prometheus support for
virtually free by using `opentelemetry-exporter-prometheus`, which will
automatically start a webserver when used with
`opentelemetry-instrument`.  When we work on configuring and starting
the worker, we can also bring in the couple lines of OTEL object setup
that would require.

Closes #11