Supports configuring and running a worker from the CLI
This ended up making me confront the fact that we didn't really deal
with how the worker knows which tasks to run.  From the client side it's
really easy, since you can just put something blindly on the docket and
we'll register it for you if it's a function.  From the worker side,
there's nowhere to really start from.

I don't think this is the final form of this, but it's a baby step.
This version supports a string path pointing to any iterable of
functions you set up, so you can have:

```
my_tasks = [task_a, task_b, task_c]
```

and then register them with `--tasks my.module:my_tasks`.  We should
really think a lot more about this.

Closes #8