Sets up `rich`, plain, and JSON logging for the worker
`rich` is doing most of the heavy lifting here for making the logging
output delightful.  I came up with a compact notation for showing the
start, end, and retry status of tasks using Unicode arrows, totally open
to feedback on that.

The plain format may need to be even plainer, but I just started there
as a starting point.

With the JSON format, I wasn't entirely sure which standard library
fields folks would want so I'm keeping it just to the level and time.

Closes #23