Adds the ability to strike and restore tasks from the docket (#34)
This adds a system of adding and removing instructions to a docket that
tell all clients and workers to ignore certain tasks.  You can specify
either all instances of a task, all tasks that have a certain argument,
or specific tasks with specific arguments.

These instructions are distributed on a separate Redis stream that all
clients monitor to receive updates.  This allows us to suppress tasks
both at scheduling and at execution.

Included are the corresponding pair of CLI commands, although they have
one limitation that they can only create parameters as strings.  We
need to come back and design a good approach to specifying other types
on the CLI.

Closes #29