Adds support for concurrent execution of tasks
The `prefetch_count` parameter has been replaced with a new
`concurrency` parameter that governs the maximum number of tasks that
are allowed to run concurrently, as well as how much if retriveed from
the Redis stream on each pass.

Big thanks to `claude-3.7-thinking` for the careful implementation.  I
did have to clean this up a good bit, but it got the framework down.

Closes #39