Runs tests in parallel (#77)
This uses `pytest-xdist` for splitting tests over multiple processes,
and coordinates the workers to create a single redis container, with
each worker using a different DB.  This allows for up to 15 workers (as
we hold off db=15 itself to be for bookkeeping of the test suite).

With some emperical testing, I found that 4 workers results in the
lowest runtime right now.

Closes #71