Adds a testing utility for self-perpetuating tasks (#78)
After a few iterations trying to get this to work with strikes, I
decided to try it with task keys and things ended up much simpler.  The
`_strike_conditions` is a nice facility that may come in handy for
some other applications in the future (not sure yet, though).  Works
like a charm for this one.

Closes #58