First-class support for self-perpetuating tasks
With the new `Perpetual` dependency, a user can request that a task
automatically perpetuate itself indefinitely.  `Perpetual` gives an
API for cancelling the chain of tasks, as well as adjusting parameters
between executions.

The typing for `perpetual.perpetuate(*args, **kwargs)` isn't aware of
the tasks' function signature, and I'm not sure how that would be
possible without requiring the user to make some pretty onerous generics
and a `Protocol` representing the task.