A first take at idempotency, which addresses tasks in the stream
Without addressing per-task locking on the task key, this makes sure
that task keys are treated idempotently between the queue and the
stream.  A key is set for each task as it is queued, and that key exists
until the task starts execution.  This helps to tie the lifetime of that
key to the redelivery mechanism, so that if a worker crashes during the
window between retrieving the message and deleting that key, it will
still be available to execute from the stream after the redelivery
timeout.