Decouples the scheduling of due tasks from the working of them (#93)
This is an optimization I've had my eye on for a bit: we don't need to
go out to the future schedule queue as frequently as we need to work
tasks on the stream.  Here, I've split that loop out separately from the
worker loop, and given it a separately configurable scheduling
resolution setting.