Allow for automatically adding `Perpetual` tasks (#97)
One major reliability issue in a "find and flood" task architecture is
if your "find" task stops getting scheduled due to errors, crashes, or
implementation mistakes.  Having a reliable facility to make sure that
`Perpetual` tasks are continually on the docket is a big boon for folks
that want to use simple `docket worker` commands without writing their
own worker entrypoint and dealing with this scheduling.

Closes #81