Fixing a myriad of race conditions (#100)
Working through the stripped down find & flood example, I noticed all
sorts of problems when running multiple workers with these quick tasks.

First, right off the bat, we had massive race condition in
`Docket.schedule` where two processes could both check if the task was
already scheduled, both see "nope", and then both proceed to schedule
the tasks.  I was seeing massive overscheduling of quick tasks.

Next, the loop for constantly adding perpetual tasks was deeply flawed
because it had no context or memory of the timing of the task's natural
rescheduling, so it would often just jump in an reschedule a task right
as it was starting.  I realized that the far more reliable thing was to
preemptively reschedule the task in the future just before executing
it, so if there was any problem after that, it was already queued for
the future.  Much safer and less polling.

Third, it occurred to me overnight that `Docket.replace` wasn't atomic,
which is a major reliability issue as well.  This prompted me to
refactor `.add`, `.replace`, `.cancel`, and `.schedule` to use pipelines
that are passed to the underlying implementations.

Closes #98