Enables simple user-defined dependency functions (#103)
This changes all dependencies to async, looking ahead to when they may
be used for things like setting up database connections, etc.
Dependencies are still "flat" functions and aren't context managers
(yet) and they can't themselves have dependencies (yet).

Future plans here are to allow for pre- and post- task dependencies
as context managers, and to allow for dependency caching and reuse so
that dependencies may themselves have dependencies (like FastAPI or
pytest fixtures).

This includes a pretty significant refactoring of the worker to flatten
and simplify things, especially in the core `_worker_loop`.  The goal
here was to make it such that `_execute` could take an `Execution`
instead of a `RedisMessage`, which will help to isolate all of the redis
task bookkeeping to the worker loop and leave task execution as
something that can focus on the Python/asyncio side of things.

First part of #96