Assorted niceties
Allows for tasks to ask for a `TaskArgument`, which illuminated the
very real possibility of dependencies failing.  So I also added the
handling for dependency failures in a way that still shows up as a task
failure and doesn't crash the worker.

I also snuck in a little testing QoL improvement to start a worker
without automatic tasks being scheduled.

Closes #109
Closes #110